third_party/WebKit/LayoutTests/virtual/feature-policy-vibrate/http/tests/feature-policy-vibrate/vibrate-allowed-by-container-policy-relocate-and-no-reload-expected.txt

CONSOLE ERROR: line 15: Navigator.vibrate() is not enabled in feature policy for this frame.
This is a testharness.js-based test.
PASS Iframe navigated itself to a new origin, vibrate is disabled by container policy.
Harness: the test ran to completion.
